added search function like nadeko
